backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.10.4
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database and vector storage
faiss-cpu==1.7.4 ; python_version < '3.12'  # Pin to Python < 3.12 for compatibility
numpy==1.26.3

# HTTP client for research
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
python-jose==3.3.0

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Type checking
mypy==1.8.0

# AI/LLM
openai==1.58.1
tavily-python==0.5.0 